Added Start Menu
The buttons all work, though the "Quit Game" button's function call to Application.Quit() is ignored within the Unity Editor. It works if you make a build.

After I wake up, I'll write animation scripts for the stars, planets, and asteroids.

The planets will slowly bob up and down.

The asteroids will rotate and travel across the screen (eventually wrapping around the screen).

The stars will slowly fade in and out to give the illusion of glimmering.

Once we agree on a game title, I'll make a stylized logo and add it to the Title Screen.