Fixed GameOverScene's Unresponsive UI
2-step solution:
1) Ensure that Image GameObjects are not listed below the Canvas GameObject in the Hierarchy panel.

Sibling Objects closer to the bottom of a list are closer to the front layer. Objects can sometimes "cover up" the sibling Objects above them ("behind" them).

In our case, an Image was blocking access to the UI Canvas.

2) When the title screen Scene transitions to the level Scene, one of the level's scripts turns the cursor invisible and locks it in place.

So I added a few lines in the Menu script to undo those cursor modifications.

Now, when the GameOverScene is loaded, the user can play again or quit the game.